backend/requirements.txt
fastapi
strawberry-graphql
sqlalchemy
uvicorn